+++ index.html
@@ -235,5 +235,6 @@
   <script type="module" src="/js/util.js"></script>
   <script type="module" src="/js/main.js"></script>
   <script type="module" src="/js/data.js"></script>
+  <script type="module" src="/js/photo-descriptions.js"></script>
 </body>
 </html>
